100 - SQLDataHolder into separate file
