+++ tel2puml/otel_to_pv/data_holders/__init__.py
@@ -1,0 +1,3 @@
+"""init file for data_holders package."""
+from .sql_data_holder.sql_dataholder import SQLDataHolder
+from .base import DataHolder
