+++ tests/tel2puml/otel_to_pv/data_sources/json_data_source/conftest.py
@@ -1,12 +1,11 @@
-"""Configuration for tests in otel_ingestion module."""
+"""Configuration for tests in json data source."""
 import pytest
 
 from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
     FieldSpec,
 )
-from tel2puml.otel_to_pv.data_sources.json_data_source.json_jq_converter import (
-    JQVariableTree,
-)
+from tel2puml.otel_to_pv.data_sources.json_data_source.json_jq_converter \
+    import JQVariableTree
 
 
 @pytest.fixture
